feat: extend visual recognition pipeline with VL review, monitoring, and ops tooling
- Backend: add keyframes API, metrics core, Celery tasks module; refine areas/cameras/edge/notifications/reviews/tasks endpoints; update event service, storage, VL service, schemas, Dockerfile, pyproject, prod compose
- Edge: refine main/config/pipeline; add edge test suite and .env.test.example
- Frontend: add Admin view, refine App/main/router/views; regenerate package-lock
- Docs: add business-flow.html and person-behavior-design.md
- Monitoring: add Prometheus/Grafana provisioning and dashboard
- Tools: add seed/validate scripts and real-test guide

+++ backend/app/schemas.py
@@ -24,7 +24,7 @@ class AreaCreate(BaseModel):
     camera_id: str
     name: str
     area_type: str = "gate"
-    polygon: list[list[int]] = []
+    polygon: list[list[float]] = []
     stay_threshold_seconds: int = 300
     high_risk_seconds: int = 600
 
@@ -32,7 +32,7 @@ class AreaCreate(BaseModel):
 class AreaUpdate(BaseModel):
     name: str | None = None
     area_type: str | None = None
-    polygon: list[list[int]] | None = None
+    polygon: list[list[float]] | None = None
     stay_threshold_seconds: int | None = None
     high_risk_seconds: int | None = None
     enabled: bool | None = None
